Migrate login pages to TS
